* develop:
  optimize code
  优化类目api
  inspect and optimise code
  release v1.21.05_0817290 1. 新增 前台类目。 2. 新增 现金授权查询功能。 3. 优化 现金授权金额消息重复问题。
  新增 点击前台类目商品收银功能。
  新增 删除类目+修改类目名称+修改商品零售价+新建商品到类目功能。
  新增 现金授权查询功能。
  新增 前台类目功能。
  新增 frontCategoryId字段到商品库表
  新增 前台菜场前台类目功能。
  修复 第二次打开支付窗口，页面元素状态没有刷新的问题。
  修复 商品数据不同步的问题。
  release v1.21.04_0812288 1. 修复 提交现金授权金额失败的问题。 2. 优化 升级bugly SDK，优化升级策略。 3. 优化 删除冗余文件资源。 4. 新增 线上预定订单打印客户姓名和手机号。 5. 新增 现金授权/清分充值后打印交易凭条。 6. 新增 手动提交现金营业额功能。 7. 新增 现金授权额度预警提示。
  优化 调整营业额支付流水同步策略。
  inspect code
  优化 删除无效或过期的csv文件
  优化打印功能； 新增 授权/清分充值单据打印。 新增 线上预定订单打印姓名and手机号
  优化 点击线上订单按钮后自动隐藏badge红点。
  修复 提交营业额现金后调用commitCash接口失败的问题。

+++ self-service-cashier/src/main/java/com/mfh/litecashier/ui/dialog/ModifyLocalCategoryDialog.java
@@ -10,20 +10,32 @@ import android.view.KeyEvent;
 import android.view.MotionEvent;
 import android.view.View;
 import android.view.WindowManager;
+import android.view.inputmethod.InputMethodManager;
 import android.widget.Button;
 import android.widget.EditText;
 import android.widget.ImageButton;
 import android.widget.TextView;
 
+import com.alibaba.fastjson.JSONObject;
 import com.bingshanguxue.cashier.database.entity.PosLocalCategoryEntity;
 import com.bingshanguxue.cashier.database.service.PosLocalCategoryService;
+import com.mfh.comn.net.data.IResponseData;
+import com.mfh.framework.MfhApplication;
+import com.mfh.framework.api.category.CateApi;
+import com.mfh.framework.api.category.ScCategoryInfoApi;
+import com.mfh.framework.core.logger.ZLogger;
 import com.mfh.framework.core.utils.DensityUtil;
 import com.mfh.framework.core.utils.DeviceUtils;
 import com.mfh.framework.core.utils.DialogUtil;
 import com.mfh.framework.core.utils.StringUtils;
-import com.mfh.framework.helper.SharedPreferencesManager;
+import com.mfh.framework.login.logic.MfhLoginService;
+import com.mfh.framework.net.NetCallBack;
+import com.mfh.framework.net.NetProcessor;
+import com.mfh.framework.network.NetWorkUtil;
 import com.mfh.framework.uikit.dialog.CommonDialog;
+import com.mfh.litecashier.CashierApp;
 import com.mfh.litecashier.R;
+import com.mfh.litecashier.service.DataSyncManager;
 
 
 /**
@@ -67,33 +79,42 @@ public class ModifyLocalCategoryDialog extends CommonDialog {
 
         tvTitle.setText("编辑栏目");
 
+
+        InputMethodManager imm = (InputMethodManager) getContext().getSystemService(Context.INPUT_METHOD_SERVICE);
+        imm.showSoftInput(etName, InputMethodManager.SHOW_IMPLICIT);
         etName.setOnTouchListener(new View.OnTouchListener() {
             @Override
             public boolean onTouch(View v, MotionEvent event) {
                 if (event.getAction() == MotionEvent.ACTION_UP) {
-                    if (SharedPreferencesManager.isSoftKeyboardEnabled()){
-                        DeviceUtils.showSoftInput(getContext(), etName);
-                    }
-                    else{
-                        DeviceUtils.hideSoftInput(getContext(), etName);
-                    }
+                    DeviceUtils.showSoftInput(getContext(), etName);
                 }
                 etName.requestFocus();
-//                etInput.setSelection(etInput.length());
+                etName.setSelection(etName.length());
                 //返回true,不再继续传递事件
                 return true;
             }
         });
+        //获取焦点后自动弹出键盘
+//        etName.setOnFocusChangeListener(new View.OnFocusChangeListener() {
+//            @Override
+//            public void onFocusChange(View v, boolean hasFocus) {
+//                if (hasFocus) {
+//                    // request keyboard// request keyboard
+//                    getWindow().setSoftInputMode(WindowManager.LayoutParams.SOFT_INPUT_STATE_ALWAYS_VISIBLE);
+//                }
+//            }
+//        });
         etName.setOnKeyListener(new EditText.OnKeyListener() {
             @Override
             public boolean onKey(View v, int keyCode, KeyEvent event) {
-//                ZLogger.d(String.format("setOnKeyListener(etQuery):keyCode=%d, action=%d", keyCode, event.getAction()));
+                ZLogger.d(String.format("setOnKeyListener(etQuery):keyCode=%d, action=%d", keyCode, event.getAction()));
                 if (keyCode == KeyEvent.KEYCODE_ENTER || keyCode == KeyEvent.KEYCODE_NUMPAD_ENTER) {
                     //按下回车键后会执行两次，
                     // 猜测一，输入框会自动捕获回车按键，自动切换焦点到下一个控件；
                     // 猜测二，通过打印日志观察发现，每次按下按键，都会监听到两次键盘事件，重复导致。
                     if (event.getAction() == MotionEvent.ACTION_UP) {
 //                        doUpdate();
+                        DeviceUtils.hideSoftInput(getContext(), etName);
                     }
                     return true;
                 }
@@ -150,7 +171,7 @@ public class ModifyLocalCategoryDialog extends CommonDialog {
 //        getWindow().setAttributes(p);
 
         //hide soft input
-        getWindow().setSoftInputMode(WindowManager.LayoutParams.SOFT_INPUT_STATE_HIDDEN);
+//        getWindow().setSoftInputMode(WindowManager.LayoutParams.SOFT_INPUT_STATE_HIDDEN);
     }
 
     public void init(PosLocalCategoryEntity categoryEntity, DialogListener dialogListener){
@@ -174,7 +195,13 @@ public class ModifyLocalCategoryDialog extends CommonDialog {
         super.show();
 
         etName.requestFocus();
-        DeviceUtils.hideSoftInput(getOwnerActivity());
+        etName.setSelection(etName.length());
+//        DeviceUtils.showSoftInput(getContext(), etName);
+//        DeviceUtils.hideSoftInput(getOwnerActivity());
+//        InputMethodManager imm = (InputMethodManager) getContext().getSystemService(Context.INPUT_METHOD_SERVICE);
+//        imm.showSoftInput(etName, InputMethodManager.SHOW_IMPLICIT);
+
+        DeviceUtils.toggleSoftInput(getContext());
     }
 
     /**
@@ -187,19 +214,62 @@ public class ModifyLocalCategoryDialog extends CommonDialog {
             return;
         }
 
-
         if (mCategoryEntity == null){
             DialogUtil.showHint("类目无效");
             return;
         }
 
-        mCategoryEntity.setName(queryText);
-        PosLocalCategoryService.get().saveOrUpdate(mCategoryEntity);
-        DialogUtil.showHint("修改成功");
-        dismiss();
-        if (listener != null){
-            listener.onComplete();
+        NetCallBack.NetTaskCallBack responseRC = new NetCallBack.NetTaskCallBack<String,
+                NetProcessor.Processor<String>>(
+                new NetProcessor.Processor<String>() {
+                    @Override
+                    protected void processFailure(Throwable t, String errMsg) {
+                        super.processFailure(t, errMsg);
+                        ZLogger.df("创建前台类目失败, " + errMsg);
+                    }
+
+                    @Override
+                    public void processResult(IResponseData rspData) {
+//                        {"code":"0","msg":"删除成功!","version":"1","data":""}
+                        //新建类目成功，保存类目信息，并触发同步。
+                        try {
+                            if (rspData == null) {
+                                return;
+                            }
+//
+//                            RspValue<String> retValue = (RspValue<String>) rspData;
+//                            String result = retValue.getValue();
+//                            Long code = Long.valueOf(result);
+
+                            mCategoryEntity.setName(queryText);
+                            PosLocalCategoryService.get().saveOrUpdate(mCategoryEntity);
+                            DialogUtil.showHint("修改成功");
+                            dismiss();
+                            if (listener != null){
+                                listener.onComplete();
+                            }
+
+                            DataSyncManager.get().sync(DataSyncManager.SYNC_STEP_FRONTEND_CATEGORY);
+                        } catch (Exception e) {
+                            ZLogger.ef(e.toString());
+                        }
+                    }
+                }
+                , String.class
+                , CashierApp.getAppContext()) {
+        };
+
+        if (!NetWorkUtil.isConnect(MfhApplication.getAppContext())){
+            DialogUtil.showHint(R.string.toast_network_error);
+            return;
         }
+
+        JSONObject jsonObject = new JSONObject();
+        jsonObject.put("id", mCategoryEntity.getId());
+        jsonObject.put("nameCn", queryText);
+        jsonObject.put("catePosition", CateApi.CATE_POSITION_FRONT);
+        jsonObject.put("tenantId", MfhLoginService.get().getSpid());
+        ScCategoryInfoApi.update(jsonObject.toJSONString(), responseRC);
     }
 
     /**
@@ -211,14 +281,52 @@ public class ModifyLocalCategoryDialog extends CommonDialog {
             return;
         }
 
+        NetCallBack.NetTaskCallBack responseRC = new NetCallBack.NetTaskCallBack<String,
+                NetProcessor.Processor<String>>(
+                new NetProcessor.Processor<String>() {
+                    @Override
+                    protected void processFailure(Throwable t, String errMsg) {
+                        super.processFailure(t, errMsg);
+                        ZLogger.df("创建前台类目失败, " + errMsg);
+                    }
 
-        PosLocalCategoryService.get().deleteById(String.valueOf(mCategoryEntity.getId()));
-        DialogUtil.showHint("删除成功");
-        dismiss();
+                    @Override
+                    public void processResult(IResponseData rspData) {
+//                        {"code":"0","msg":"删除成功!","version":"1","data":""}
+                        //新建类目成功，保存类目信息，并触发同步。
+                        try {
+                            if (rspData == null) {
+                                return;
+                            }
+//
+//                            RspValue<String> retValue = (RspValue<String>) rspData;
+//                            String result = retValue.getValue();
+//                            Long code = Long.valueOf(result);
 
-        if (listener != null){
-            listener.onComplete();
+
+                            PosLocalCategoryService.get().deleteById(String.valueOf(mCategoryEntity.getId()));
+                            DialogUtil.showHint("删除成功");
+                            dismiss();
+
+                            if (listener != null){
+                                listener.onComplete();
+                            }
+
+                            DataSyncManager.get().sync(DataSyncManager.SYNC_STEP_FRONTEND_CATEGORY);
+                        } catch (Exception e) {
+                            ZLogger.ef(e.toString());
+                        }
+                    }
+                }
+                , String.class
+                , CashierApp.getAppContext()) {
+        };
+
+        if (!NetWorkUtil.isConnect(MfhApplication.getAppContext())){
+            DialogUtil.showHint(R.string.toast_network_error);
+            return;
         }
-    }
 
+        ScCategoryInfoApi.delete(mCategoryEntity.getId(), responseRC);
+    }
 }
